Major Update: Resized all the forms and modified the report generating feature, and also added an Icon
