+++ Program.cs
@@ -22,17 +22,7 @@ namespace Warehouse_System
             string projectFolder = System.IO.Path.GetFullPath(System.IO.Path.Combine(Application.StartupPath, @"..\..\"));
             AppDomain.CurrentDomain.SetData("DataDirectory", projectFolder);
 
-            //Application.Run(new LoginForm());
-            //Application.Run(new BranchForm());
-            //Application.Run(new AccessoryForm());
-            //Application.Run(new SupplierForm());
-            //Application.Run(new ProductForm());
-            Application.Run(new ProductRestockForm());
-            //Application.Run(new DispatchForm());
-            //Application.Run(new AdminUI());
-            //Application.Run(new WarehouseManagerUI());
-            //Application.Run(new WarehouseStaffUI());
-            //Application.Run(new EmployeeForm());
+            Application.Run(new LoginForm());
         }
     }
 }
